bug/keyp-193 : move onRemoteReaderEvent to ObservableVirtualReader
