+++ java/component/keyple-plugin/remotese/src/main/java/org/eclipse/keyple/plugin/remotese/pluginse/VirtualObservableReaderImpl.java
@@ -11,10 +11,12 @@
  ********************************************************************************/
 package org.eclipse.keyple.plugin.remotese.pluginse;
 
+import static org.eclipse.keyple.core.seproxy.ChannelControl.CLOSE_AFTER;
 import java.util.*;
 import org.eclipse.keyple.core.seproxy.event.AbstractDefaultSelectionsRequest;
 import org.eclipse.keyple.core.seproxy.event.ObservableReader;
 import org.eclipse.keyple.core.seproxy.event.ReaderEvent;
+import org.eclipse.keyple.core.seproxy.exception.KeypleReaderException;
 import org.eclipse.keyple.core.seproxy.protocol.TransmissionMode;
 import org.eclipse.keyple.plugin.remotese.exception.KeypleRemoteException;
 import org.eclipse.keyple.plugin.remotese.pluginse.method.RmSetDefaultSelectionRequestTx;
@@ -23,7 +25,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * Observable Virtual Reader add Observable methods to VirtualReaderImpl
+ * Observable Virtual Reader
+ *
+ * add Observable methods to VirtualReaderImpl
  */
 final class VirtualObservableReaderImpl extends VirtualReaderImpl
         implements VirtualObservableReader {
@@ -110,21 +114,63 @@ final class VirtualObservableReaderImpl extends VirtualReaderImpl
     }
 
     @Override
-    public void notifySeProcessed() {
-        // TODO Check why this method is empty here
+    public final void notifySeProcessed() {
+        if (forceClosing) {
+            try {
+                // close the physical channel thanks to CLOSE_AFTER flag
+                processSeRequest(null, CLOSE_AFTER);
+                logger.trace(
+                        "Explicit communication closing requested, starting removal sequence.");
+            } catch (KeypleReaderException e) {
+                logger.error("KeypleReaderException while terminating. {}", e.getMessage());
+            }
+        } else {
+            logger.trace("Explicit physical channel closing already requested.");
+        }
     }
 
     @Override
     public void startSeDetection(PollingMode pollingMode) {
-        logger.error(
-                "startSeDetection is not implemented in VirtualObservableReaderImpl, please use the local method");
+        logger.warn(
+                "startSeDetection is not implemented in VirtualObservableReaderImpl, please use the method on the slave node");
     }
 
     @Override
     public void stopSeDetection() {
-        logger.error(
-                "stopSeDetection is not implemented in VirtualObservableReaderImpl, please use the local method");
+        logger.warn(
+                "stopSeDetection is not implemented in VirtualObservableReaderImpl, please use the method on the slave node");
     }
+
+    /*
+     * PACKAGE PRIVATE
+     */
+
+    /**
+     * When an event occurs on the Remote LocalReader, notify Observers
+     *
+     * @param event
+     */
+    void onRemoteReaderEvent(final ReaderEvent event) {
+        final VirtualReaderImpl thisReader = this;
+
+        logger.debug("{} EVENT {} ", this.getName(), event.getEventType());
+
+        if (((VirtualObservableReader) thisReader).countObservers() > 0) {
+            // launch event another thread to permit blocking method to be used in update
+            // method (such as transmit)
+            executorService.execute(new Runnable() {
+                @Override
+                public void run() {
+                    ((VirtualObservableReader) thisReader).notifyObservers(event);
+                }
+            });
+        } else {
+            logger.debug(
+                    "An event was received but no observers are declared into VirtualReader : {} {}",
+                    thisReader.getName(), event.getEventType());
+        }
+    }
+
 
     @Override
     public void setDefaultSelectionRequest(
