+++ java/component/keyple-plugin/remotese/src/test/java/org/eclipse/keyple/plugin/remotese/integration/VirtualReaderEventTest.java
@@ -184,9 +184,6 @@ public class VirtualReaderEventTest extends VirtualReaderBaseTest {
         virtualReader.removeObserver(obs);
 
         Assert.assertEquals(0, lock.getCount());
-
-        // https://github.com/calypsonet/keyple-java/issues/420
-        // Assert.assertEquals(0, masterAPI.getPlugin().getReaders().size());
     }
